Use only filename part instead of full url to pick new hls fragments
The same fragment might have different urls on two consequent updates of
hls playlist, in which case last_fragment exact value will not be
found in the urls list, which will lead to new_idx being set to zero and
already processed fragments being downloaded for a second time.

To deal with this instead of full fragments urls only filename parts are
used for comparison, since they only contain fragment number, which is
unique for a given fragment and doesn't change.